[新增功能]: 新增消息开始监听和结束监听的逻辑
content:
1.新增监听looper消息的逻辑

+++ somersaultCloud/src/main/java/org/somersault/cloud/lib/core/anr/bean/MonitorMessage.kt
@@ -15,4 +15,4 @@ import java.io.Serializable
  * 修订历史：
  * ================================================
  */
-data class MonitorMessage(val handlerName:String,val handlerAddress : String,val callBackName:String,val messageWhat : Int ,var msgId : Int = 0) : Serializable
+data class MonitorMessage(val handlerName:String,val handlerAddress : String,val callBackName:String,val messageWhat : Int ,var msgId : Long = 0L) : Serializable
